+++ app/controllers/web_client_controller.rb
@@ -1,5 +1,6 @@
 class WebClientController < ApplicationController
   before_action :set_web_client
+  before_action :set_contact, only: [:add_another, :ignore_another]
 
   def show
   end
@@ -25,8 +26,13 @@ class WebClientController < ApplicationController
   end
 
   def add_another
-    WebClient::AddContact.new(Contact.find(params[:contact_id])).do
+    WebClient::AddContact.new(@contact).do
     render json: { status: 'feature in progress' }
+  end
+
+  def ignore_another
+    WebClient::IgnoreContact.new(@contact).do
+    redirect_to web_client_path
   end
 
   private
@@ -38,5 +44,9 @@ class WebClientController < ApplicationController
   def set_web_client
     @web_client = WebClientDecorator.decorate(WebClient::ActionHandler.new(params[:id]))
     raise WebClient::NotFound unless @web_client.valid?
+  end
+
+  def set_contact
+    @contact = Contact.find(params[:contact_id])
   end
 end