+++ app/models/contact.rb
@@ -23,10 +23,18 @@ class Contact < ActiveRecord::Base
   scope :order_by_score, -> { order('total_score DESC') }
   scope :expired, -> { where('expires_at < ?', Time.now) }
 
-  scope :not_proposed, -> { includes(:notifications).where(notifications: { id: nil }) }
   scope :not_equals_with_owner, -> { where('contacts.owner_mkey != contacts.zazo_mkey OR contacts.zazo_mkey IS NULL') }
-  scope :not_friends_with_owner, -> { not_equals_with_owner.where("contacts.additions->>'marked_as_friend' = 'false'") }
-  scope :friends_with_owner, -> { not_equals_with_owner.where("contacts.additions->>'marked_as_friend' = 'true'") }
+
+  scope :proposed, -> { with_notifications.where.not(notifications: { id: nil }) }
+  scope :not_proposed, -> { with_notifications.where(notifications: { id: nil }) }
+
+  scope :not_friends, -> { not_equals_with_owner.where("contacts.additions->>'marked_as_friend' = 'false'") }
+  scope :friends, -> { not_equals_with_owner.where("contacts.additions->>'marked_as_friend' = 'true'") }
+
+  scope :rejected, -> { where("contacts.additions->>'rejected_by_owner' = 'true'") }
+  scope :not_rejected, -> { where("(contacts.additions->>'rejected_by_owner') IS NULL") }
+
+  scope :suggestible, -> { not_rejected.not_proposed.not_friends }
 
   before_save { self.expires_at = 5.days.from_now }