+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### v0.5.0
 - :bulb: Added status field for contacts concepts
 - :bulb: Implemented filter system for contacts concepts
+- :bulb: Implemented feature to ignore any contact on web-client
 - :arrow_up: Updated web-client navbar
 - :arrow_up: Updated web-client notification messages
 - :hammer: Fixed dockerfile assets precompile